Fixed compile error in debug output of class PivotTable
The compiler error occurs when the DEBUG REPORTS macro is set.

CCBUG:387040

+++ kmymoney/reports/pivottable.cpp
@@ -1022,7 +1022,7 @@ void PivotTable::convertToDeepCurrency()
             it_row.value()[ePrice][column] = PivotCell(priceValue.convert(10000));
           }
 
-          DEBUG_OUTPUT_IF(conversionfactor != MyMoneyMoney::ONE , QString("Factor of %1, value was %2, now %3").arg(conversionfactor).arg(DEBUG_SENSITIVE(oldval)).arg(DEBUG_SENSITIVE(it_row.value()[eActual][column].toDouble())));
+          DEBUG_OUTPUT_IF(conversionfactor != MyMoneyMoney::ONE , QString("Factor of %1, value was %2, now %3").arg(conversionfactor.toDouble()).arg(DEBUG_SENSITIVE(oldval.toDouble())).arg(DEBUG_SENSITIVE(it_row.value()[eActual][column].toDouble())));
 
           ++column;
         }
